refactor: 잘못된 경로에서 쓰던 page 삭제, 디자인과 기능만 맞는 경로로 마이그레이션
